refactor: auth그륩 페이지 rcc/rsc 폴더 분리 (#180)
