+++ src/app/(auth)/_components/client/auth-layout.tsx
