+++ src/app/(mypage)/account/_components/account-security-info.tsx
@@ -5,7 +5,6 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
-import PasswordInput from '@/components/features/auth/auth-password-input';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
 import { ERROR_MESSAGES } from '@/constants/mypage.constants';
@@ -13,6 +12,7 @@ import { fetchUpdatePassword } from '@/lib/apis/profile/update-profile.api';
 import useCustomToast from '@/lib/hooks/use-custom-toast';
 import useProviderFromCookie from '@/lib/hooks/use-get-provider';
 import { changePasswordSchema } from '@/lib/schemas/change-password-schema';
+import PasswordInput from '@/components/features/input/password-input';
 
 const SECURITY_INFO_STYLE = {
   container:
